refactor: improve speed of recipe loading
Makes use of suspense for loading on top of existing loading componet.

This also removes the generateMetaData. This is partially done becuse it
did not actually work as it seems, and as it for now is slowing down the
load time.

When at some point introducing metadata again, we need to ensure it does
not impact loading speed.